+++ packages/environment/src/exchanges.ts
@@ -1,6 +1,7 @@
 export enum Exchange {
   ONE_INCH_V5 = "ONE_INCH_V5",
-  PARASWAP = "PARASWAP",
+  PARASWAP_V5 = "PARASWAP_V5",
+  PARASWAP_V6 = "PARASWAP_V6",
   PMM_2 = "PMM_2",
   PMM_2_KYC = "PMM_2_KYC",
   UNISWAP_V2 = "UNISWAP_V2",