+++ packages/environment/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @enzymefinance/environment
 
+## 6.3.34
+
+### Patch Changes
+
+- [#696](https://github.com/enzymefinance/sdk/pull/696) [`a80ec5d`](https://github.com/enzymefinance/sdk/commit/a80ec5ddb0dcaf8f3ea7a81c5a0f245b1fa5f6a0) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add RSUP
+
+## 6.3.33
+
+### Patch Changes
+
+- [#663](https://github.com/enzymefinance/sdk/pull/663) [`898a8b8`](https://github.com/enzymefinance/sdk/commit/898a8b825d88537402e94fd559748e52218787ac) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add Pendle V2 assets
+
+## 6.3.32
+
+### Patch Changes
+
+- [#689](https://github.com/enzymefinance/sdk/pull/689) [`e525c38`](https://github.com/enzymefinance/sdk/commit/e525c38fb88892690c9b9337a763b90fce32a125) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Fix ParaSwapV6.2 AugustusSwapperAddress
+
+## 6.3.31
+
+### Patch Changes
+
+- [#687](https://github.com/enzymefinance/sdk/pull/687) [`60ad801`](https://github.com/enzymefinance/sdk/commit/60ad8019153723a32d56f4796d799d6120b48397) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Add ParaSwapV6AugustusSwapper to external contracts
+
+## 6.3.30
+
+### Patch Changes
+
+- [#682](https://github.com/enzymefinance/sdk/pull/682) [`df7cd76`](https://github.com/enzymefinance/sdk/commit/df7cd76a011ec24976be7e4e4b008541452548f4) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Base adjustments
+
+## 6.3.29
+
+### Patch Changes
+
+- [#659](https://github.com/enzymefinance/sdk/pull/659) [`76dc58e`](https://github.com/enzymefinance/sdk/commit/76dc58e33b617d3577959245179c9762a8ce11f2) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add single asset deposit queue
+
+- Updated dependencies [[`76dc58e`](https://github.com/enzymefinance/sdk/commit/76dc58e33b617d3577959245179c9762a8ce11f2)]:
+  - @enzymefinance/sdk@1.2.11
+
+## 6.3.28
+
+### Patch Changes
+
+- [#679](https://github.com/enzymefinance/sdk/pull/679) [`176dd2a`](https://github.com/enzymefinance/sdk/commit/176dd2a4768a9332b88edd8d629b6bcfc45ed634) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Add price feed factories
+
+- [#680](https://github.com/enzymefinance/sdk/pull/680) [`adf31fd`](https://github.com/enzymefinance/sdk/commit/adf31fdddcbe47200b0acc96423d8a41c9ae9275) Thanks [@KedziaPawel](https://github.com/KedziaPawel)! - Remove liquity
+
+- Updated dependencies [[`adf31fd`](https://github.com/enzymefinance/sdk/commit/adf31fdddcbe47200b0acc96423d8a41c9ae9275)]:
+  - @enzymefinance/sdk@1.2.10
+
+## 6.3.27
+
+### Patch Changes
+
+- [#658](https://github.com/enzymefinance/sdk/pull/658) [`4a06def`](https://github.com/enzymefinance/sdk/commit/4a06def2ddded3ad3a69fc909c89ad4d89b5cdd2) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Add ParaSwapV6 Integration
+
+- Updated dependencies [[`4a06def`](https://github.com/enzymefinance/sdk/commit/4a06def2ddded3ad3a69fc909c89ad4d89b5cdd2)]:
+  - @enzymefinance/sdk@1.2.9
+
 ## 6.3.26
 
 ### Patch Changes